Removed unused files and moved Connection class
